requirements.txt
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
gradio>=4.44.0
huggingface-hub>=0.24.0
numpy>=1.26.0
python-dotenv>=1.0.0